Got basic SWIG generation working.
By using the base Redland.i interface file that ships with Redland,
one can now generate the Redland_wrap.c and compile it with R. Still
need to work out issues with why the .i file needed to be modified
for a couple of functions. Also, upgraded to using swig3.0 in the
Dockerfile.