Automatically modify the generated Redland.R file.
Redland.R is generated by SWIG, but it is missing two class definitions.  I manually added these to Redland.R.in and now append that onto Redland.R after SWIG has generated it.  This allows Redland.R to load and tests to pass, but is obviously obscuring a bug in generation which needs to be fixed.