Added function rasqal_world_set_warning_level, with associated C bindings.
The function rasqal_world_set_warning_level is used to set the warning
level of the rasqal library, which is defaults to fairly verbose. The
function is now part of the redland-bindings R library, and can be
called with a World object and an integer warning level. The level is
now set to 10 (just errors, no warnings) on the default World object.